Move file access checks to own helpers file
